refactor: #7 프로토콜 이동
