+++ Recipe/Protocols/Protocols.swift
@@ -5,4 +5,39 @@
 //  Created by KindSoft on 2023/07/03.
 //
 
-import Foundation
+import UIKit
+
+/// For Coordinator
+protocol GoCoordinatorProtocol: Coordinator {
+    func startReadyFlow()
+}
+
+protocol SplashCoordinatorProtocol: Coordinator {
+    func showSplashView()
+}
+
+protocol LoginCoordinatorProtocol: Coordinator {
+    func showLoginViewController()
+}
+
+protocol RegisterCoordinatorProtocol: Coordinator {
+    func showRegisterViewController()
+}
+
+protocol HomeCoordinatorProtocol: Coordinator {
+    func startReadyFlow()
+}
+
+/// Else
+protocol RefrigeratorDetailDelegate{
+    func onClickButton(_ senderTitle: String)
+}
+
+protocol PriceTrendHeaderDelegate {
+    func showAllData()
+}
+
+protocol TagCellDelegate {
+    func deleteButtonTapped(sender: Int)
+}
+
